backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.0.0
mysqlclient==2.1.1
python-dotenv==1.0.0
Pillow==9.5.0 